spec: menu/radio: allow values to have any String
I've seen so many modules demand special codewords -- strptime()
formats, format() formats, JSON paths, and so on.

The only reason we restricted the value formats in the past was to keep
our options open. We were thinking, "maybe a future module API would let
users use variable names?" Now we have more information, and we know
that allowing any value is a better way to go.